quizListが共有されてしまう問題を修正
