+++ packages/frontend/src/components/QuizList.tsx
@@ -1,9 +1,10 @@
 import { Quiz } from "@/types"
-import { useMylistInfomations } from "@/hooks/useMylists";
+import { useMylists } from "@/hooks/useMylists";
 import QuizCard from "./QuizCard"
+import { Center, Loader } from "@mantine/core";
 
 interface QuizListProps {
-  quizzes: Quiz[],
+  quizzes?: Quiz[],
   isHidden?: boolean,
   coloring?: boolean,
 }
@@ -13,21 +14,34 @@ export default function QuizList({
   isHidden = false,
   coloring = false,
 }: QuizListProps) {
-  const { mylists } = useMylistInfomations();
+  const { mylists } = useMylists();
 
   return (
     <>
-      {quizzes?.map((quiz, idx) => (
-        <QuizCard 
-          key={`${idx}${isHidden}`}
-          index={idx+1}
-          quiz={quiz}
-          mylists={mylists || []}
-          coloring={coloring}
-          isHidden={isHidden}
-          mb={10}
-        />
-      ))}
+    {
+      !!quizzes 
+      ?
+      <>
+      {
+        quizzes.map((quiz, idx) => (
+          <QuizCard 
+            key={`${idx}${isHidden}`}
+            index={idx+1}
+            quiz={quiz}
+            mylists={mylists || []}
+            coloring={coloring}
+            isHidden={isHidden}
+            mb={10}
+          />
+        ))
+      }
+      { quizzes.length == 0 ? <Center>No data</Center> : null }
+      </>
+      : 
+      <Center>
+        <Loader variant="dots"/>
+      </Center>
+    }
     </>
   )
 }
