+++ packages/frontend/src/pages/search.tsx
@@ -1,103 +1,17 @@
-import QuizList from '../components/QuizList'
-import QuizControllBar from '../components/QuizControllBar'
-import useQuizzes from '../hooks/useQuizzes'
-import { Center, Grid, Group, Loader } from '@mantine/core'
-import FilteringModal from '../components/FilteringModal'
-import { KeywordOption } from '../types'
-import { useState } from 'react'
-import QuizPagination from '../components/QuizPagination'
-import QuizShuffleButton from '../components/QuizShuffleButton'
-import QuizHiddenAnswerButton from '../components/QuizHiddenAnswerButton'
+import { useEffect } from 'react';
+import QuizViewer from '@/components/QuizViewer';
+import useQuizzes from '@/hooks/useQuizzes';
 
 export default function Search() {
-  const [activePage, setPage] = useState(1);
-  const [ isHidden, setIsHidden ] = useState(false);
-  const { quizzes, params, setParams } = useQuizzes()
-
-  const size = !!quizzes && !!quizzes.length ? quizzes[0].size : 0
-
-  const toFilter = (
-    workbooks?: string[], 
-    levels?: string[], 
-    keyword?: string, 
-    keywordOption?: KeywordOption,
-    perPage?: number,
-  ) => {
-    setPage(1)
-    setParams({ 
-      ...params, 
-      page: 1, 
-      seed: undefined,
-      perPage,
-      workbooks, 
-      levels, 
-      keyword, 
-      keywordOption,
-    })
-  }
-
-  const toShuffle = (
-    seed: number
-  ) => {
-    setPage(1)
-    setParams({
-      ...params,
-      page: 1,
-      seed
-    })
-  }
-
-  const changePage = (
-    page: number
-  ) => {
-    setPage(page)
-    setParams({...params, page})
-  }
+  const { setParams } = useQuizzes();
+  
+  useEffect(() => {
+    setParams({ perPage: 100 })
+  }, []);
 
   return (
     <>
-      <QuizControllBar
-        height={!!quizzes ? 110 : 60}
-        total={size}
-        buttons={
-          <Group>
-            <FilteringModal
-              apply={toFilter}
-            />
-            <QuizShuffleButton
-              apply={toShuffle}
-            />
-            <QuizHiddenAnswerButton
-              isHidden={isHidden}
-              onToggle={setIsHidden}
-            />
-          </Group>
-        }
-        pagination={
-          <Grid.Col mb={5}>
-            <Center>
-              <QuizPagination
-                page={activePage}
-                total={!!params?.perPage ? Math.ceil(size / params.perPage) : 0}
-                setPage={changePage}
-              />
-            </Center>
-          </Grid.Col>
-        }
-      />
-      {!!quizzes ? 
-        <>
-          <QuizList
-            quizzes={quizzes}
-            isHidden={isHidden}
-          /> 
-          { quizzes.length == 0 ? <Center>No data</Center> : null }
-        </>
-      : 
-        <Center>
-          <Loader variant="dots"/>
-        </Center>
-      }
+      <QuizViewer />
     </>
   )
 }
